Move dummy viewmodel into adapter layer
