+++ OTAnalytics/adapter_ui/dummy_viewmodel.py
